+++ src/functions_framework/event_conversion.py
@@ -13,9 +13,11 @@
 # limitations under the License.
 import re
 
-from typing import Tuple
+from datetime import datetime
+from typing import Any, Optional, Tuple
 
-from cloudevents.http import CloudEvent
+from cloudevents.exceptions import MissingRequiredFields
+from cloudevents.http import CloudEvent, from_http, is_binary
 
 from functions_framework.background_event import BackgroundEvent
 from functions_framework.exceptions import EventConversionException
@@ -47,6 +49,19 @@ _BACKGROUND_TO_CE_TYPE = {
     "providers/cloud.storage/eventTypes/object.change": "google.cloud.storage.object.v1.finalized",
 }
 
+# _BACKGROUND_TO_CE_TYPE contains duplicate values for some keys. This set contains the duplicates
+# that should be dropped when generating the inverse mapping _CE_TO_BACKGROUND_TYPE
+_NONINVERTALBE_CE_TYPES = {
+    "providers/cloud.pubsub/eventTypes/topic.publish",
+    "providers/cloud.storage/eventTypes/object.change",
+}
+
+# Maps CloudEvent types to the equivalent background/legacy event types (inverse
+# of _BACKGROUND_TO_CE_TYPE)
+_CE_TO_BACKGROUND_TYPE = {
+    v: k for k, v in _BACKGROUND_TO_CE_TYPE.items() if k not in _NONINVERTALBE_CE_TYPES
+}
+
 # CloudEvent service names.
 _FIREBASE_AUTH_CE_SERVICE = "firebaseauth.googleapis.com"
 _FIREBASE_CE_SERVICE = "firebase.googleapis.com"
@@ -54,6 +69,12 @@ _FIREBASE_DB_CE_SERVICE = "firebasedatabase.googleapis.com"
 _FIRESTORE_CE_SERVICE = "firestore.googleapis.com"
 _PUBSUB_CE_SERVICE = "pubsub.googleapis.com"
 _STORAGE_CE_SERVICE = "storage.googleapis.com"
+
+# Raw pubsub types
+_PUBSUB_EVENT_TYPE = "google.pubsub.topic.publish"
+_PUBSUB_MESSAGE_TYPE = "type.googleapis.com/google.pubsub.v1.PubsubMessage"
+
+_PUBSUB_TOPIC_REQUEST_PATH = re.compile(r"projects\/[^/?]+\/topics\/[^/?]+")
 
 # Maps background event services to their equivalent CloudEvent services.
 _SERVICE_BACKGROUND_TO_CE = {
@@ -86,11 +107,16 @@ _FIREBASE_AUTH_METADATA_FIELDS_BACKGROUND_TO_CE = {
     "createdAt": "createTime",
     "lastSignedInAt": "lastSignInTime",
 }
+# Maps Firebase Auth CloudEvent metadata field names to their equivalent
+# background event field names (inverse of _FIREBASE_AUTH_METADATA_FIELDS_BACKGROUND_TO_CE).
+_FIREBASE_AUTH_METADATA_FIELDS_CE_TO_BACKGROUND = {
+    v: k for k, v in _FIREBASE_AUTH_METADATA_FIELDS_BACKGROUND_TO_CE.items()
+}
 
 
 def background_event_to_cloudevent(request) -> CloudEvent:
-    """Converts a background event represented by the given HTTP request into a CloudEvent. """
-    event_data = request.get_json()
+    """Converts a background event represented by the given HTTP request into a CloudEvent."""
+    event_data = marshal_background_event_data(request)
     if not event_data:
         raise EventConversionException("Failed to parse JSON")
 
@@ -136,6 +162,74 @@ def background_event_to_cloudevent(request) -> CloudEvent:
     return CloudEvent(metadata, data)
 
 
+def is_convertable_cloudevent(request) -> bool:
+    """Is the given request a known CloudEvent that can be converted to background event."""
+    if is_binary(request.headers):
+        event_type = request.headers.get("ce-type")
+        event_source = request.headers.get("ce-source")
+        return (
+            event_source is not None
+            and event_type is not None
+            and event_type in _CE_TO_BACKGROUND_TYPE
+        )
+    return False
+
+
+def _split_ce_source(source) -> Tuple[str, str]:
+    """Splits a CloudEvent source string into resource and subject components."""
+    regex = re.compile(r"\/\/([^/]+)\/(.+)")
+    match = regex.fullmatch(source)
+    if not match:
+        raise EventConversionException("Unexpected CloudEvent source.")
+
+    return match.group(1), match.group(2)
+
+
+def cloudevent_to_background_event(request) -> Tuple[Any, Context]:
+    """Converts a background event represented by the given HTTP request into a CloudEvent."""
+    try:
+        event = from_http(request.headers, request.get_data())
+        data = event.data
+        service, name = _split_ce_source(event["source"])
+
+        if event["type"] not in _CE_TO_BACKGROUND_TYPE:
+            raise EventConversionException(
+                f'Unable to find background event equivalent type for "{event["type"]}"'
+            )
+
+        if service == _PUBSUB_CE_SERVICE:
+            resource = {"service": service, "name": name, "type": _PUBSUB_MESSAGE_TYPE}
+            if "message" in data:
+                data = data["message"]
+        elif service == _FIREBASE_AUTH_CE_SERVICE:
+            resource = name
+            if "metadata" in data:
+                for old, new in _FIREBASE_AUTH_METADATA_FIELDS_CE_TO_BACKGROUND.items():
+                    if old in data["metadata"]:
+                        data["metadata"][new] = data["metadata"][old]
+                        del data["metadata"][old]
+        elif service == _STORAGE_CE_SERVICE:
+            resource = {
+                "name": f"{name}/{event['subject']}",
+                "service": service,
+                "type": data["kind"],
+            }
+        else:
+            resource = f"{name}/{event['subject']}"
+
+        context = Context(
+            eventId=event["id"],
+            timestamp=event["time"],
+            eventType=_CE_TO_BACKGROUND_TYPE[event["type"]],
+            resource=resource,
+        )
+        return (data, context)
+    except (AttributeError, KeyError, TypeError, MissingRequiredFields):
+        raise EventConversionException(
+            "Failed to convert CloudEvent to BackgroundEvent."
+        )
+
+
 def _split_resource(context: Context) -> Tuple[str, str, str]:
     """Splits a background event's resource into a CloudEvent service, resource, and subject."""
     service = ""
@@ -168,3 +262,56 @@ def _split_resource(context: Context) -> Tuple[str, str, str]:
         raise EventConversionException("Resource regex did not match")
 
     return service, match.group(1), match.group(2)
+
+
+def marshal_background_event_data(request):
+    """Marshal the request body of a raw Pub/Sub HTTP request into the schema that is expected of
+    a background event"""
+    try:
+        request_data = request.get_json()
+        if not _is_raw_pubsub_payload(request_data):
+            # If this in not a raw Pub/Sub request, return the unaltered request data.
+            return request_data
+        return {
+            "context": {
+                "eventId": request_data["message"]["messageId"],
+                "timestamp": request_data["message"].get(
+                    "publishTime", datetime.utcnow().isoformat() + "Z"
+                ),
+                "eventType": _PUBSUB_EVENT_TYPE,
+                "resource": {
+                    "service": _PUBSUB_CE_SERVICE,
+                    "type": _PUBSUB_MESSAGE_TYPE,
+                    "name": _parse_pubsub_topic(request.path),
+                },
+            },
+            "data": {
+                "@type": _PUBSUB_MESSAGE_TYPE,
+                "data": request_data["message"]["data"],
+                "attributes": request_data["message"]["attributes"],
+            },
+        }
+    except (AttributeError, KeyError, TypeError):
+        raise EventConversionException("Failed to convert Pub/Sub payload to event")
+
+
+def _is_raw_pubsub_payload(request_data) -> bool:
+    """Does the given request body match the schema of a unmarshalled Pub/Sub request"""
+    return (
+        request_data is not None
+        and "context" not in request_data
+        and "subscription" in request_data
+        and "message" in request_data
+        and "data" in request_data["message"]
+        and "messageId" in request_data["message"]
+    )
+
+
+def _parse_pubsub_topic(request_path) -> Optional[str]:
+    match = _PUBSUB_TOPIC_REQUEST_PATH.search(request_path)
+    if match:
+        return match.group(0)
+    else:
+        # It is possible to configure a Pub/Sub subscription to push directly to this function
+        # without passing the topic name in the URL path.
+        return ""