+++ CHANGELOG.md
@@ -6,7 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [2.1.2] - 2020-02-23
+## [2.2.1] - 2021-06-01
+### Changed
+- Update GCF Python 3.7 backwards-compatible logging ([#131])
+
+## [2.2.0] - 2021-05-24
+### Added
+- Relax constraint to `flask<3.0` and `click<9.0` ([#129])
+
+## [2.1.3] - 2021-04-23
+### Changed
+- Change gunicorn loglevel to error ([#122])
+
+### Added
+- Add support for background to CloudEvent conversion ([#116])
+
+## [2.1.2] - 2021-02-23
 ### Added
 - Add crash header to 500 responses ([#114])
 
@@ -94,7 +109,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/GoogleCloudPlatform/functions-framework-python/compare/v2.1.2...HEAD
+[Unreleased]: https://github.com/GoogleCloudPlatform/functions-framework-python/compare/v2.2.1...HEAD
+[2.2.1]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v2.2.1
+[2.2.0]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v2.2.0
+[2.1.3]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v2.1.3
 [2.1.2]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v2.1.2
 [2.1.1]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v2.1.1
 [2.1.0]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v2.1.0
@@ -113,6 +131,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.0.1]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v1.0.1
 [1.0.0]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v1.0.0
 
+[#131]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/131
+[#129]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/129
+[#122]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/122
+[#116]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/116
 [#114]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/114
 [#107]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/107
 [#105]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/105